Split Condor tarball packaging tests
